Add TransUNet.

+++ contrib/MedicalSeg/test_tipc/prepare.sh
@@ -19,9 +19,17 @@ trainer_list=$(func_parser_value "${lines[14]}")
 
 # MODE be one of ['lite_train_lite_infer']
 if [ ${MODE} = "lite_train_lite_infer" ];then
-    mkdir -p ./test_tipc/data
-    rm -rf ./test_tipc/data/mini_levir_dataset
-    cd ./test_tipc/data/
-    wget https://bj.bcebos.com/paddleseg/dataset/mini_brainT_dataset.zip
-    unzip mini_brainT_dataset.zip  && cd ../../
+   if [ ${model_name} == "UNETR" ]; then
+      mkdir -p ./test_tipc/data
+      rm -rf ./test_tipc/data/mini_levir_dataset
+      cd ./test_tipc/data/
+      wget https://bj.bcebos.com/paddleseg/dataset/mini_brainT_dataset.zip
+      unzip mini_brainT_dataset.zip  && cd ../../
+    elif [ ${model_name} == "TransUNet" ]; then
+      rm -rf ./test_tipc/data/mini_synapse_dataset
+      cd ./test_tipc/data/
+      unzip mini_synapse_dataset.zip  && cd ../../
+    else
+        echo "Not added into TIPC yet."
+    fi
 fi
